requirements.txt
###### Requirements without Version Specifiers ######
pandas
numpy
requests
PIL
io
tweepy
json
os
seaborn
matplotlib.pyplot